Implemented data selector switcher for actions
