+++ src/Form/AddConditionForm.php
@@ -19,6 +19,8 @@ use Symfony\Component\DependencyInjection\ContainerInterface;
  */
 class AddConditionForm extends FormBase {
 
+  use ContextFormTrait;
+
   /**
    * The condition plugin manager.
    *
@@ -73,10 +75,10 @@ class AddConditionForm extends FormBase {
       return $form;
     }
 
+    // Step 2 of the form.
     /** @var \Drupal\rules\Core\RulesConditionInterface $condition */
     $condition = $this->conditionManager->createInstance($condition_name);
 
-    // Step 2 of the form.
     $form['summary'] = [
       '#markup' => $condition->summary(),
     ];
@@ -99,58 +101,6 @@ class AddConditionForm extends FormBase {
     ];
 
     return $form;
-  }
-
-  /**
-   * Provides the form part for a context parameter.
-   */
-  public function buildContextForm(array $form, FormStateInterface $form_state, $context_name, $context_definition) {
-    $form['context'][$context_name] = [
-      '#type' => 'fieldset',
-      '#title' => $context_definition->getLabel(),
-    ];
-    $form['context'][$context_name]['description'] = [
-      '#markup' => $context_definition->getDescription(),
-    ];
-
-    $mode = $form_state->get('context_' . $context_name, 'input');
-    $title = $mode == 'selector' ? $this->t('Data selector') : $this->t('Value');
-    // @todo get a description for possible values that can be filled in.
-    $description = $mode == 'selector'
-      ? $this->t("The data selector helps you drill down into the data available to Rules. <em>To make entity fields appear in the data selector, you may have to use the condition 'entity has field' (or 'content is of type').</em> More useful tips about data selection is available in <a href=':url'>the online documentation</a>.", [
-        ':url' => 'https://www.drupal.org/node/1300042',
-      ]) : '';
-    $form['context'][$context_name]['setting'] = [
-      '#type' => 'textfield',
-      '#title' => $title,
-      '#description' => $description,
-      '#required' => $context_definition->isRequired(),
-    ];
-
-    $value = $mode == 'selector' ? $this->t('Switch to the direct input mode') : $this->t('Switch to data selection');
-    $form['context'][$context_name]['switch_button'] = [
-      '#type' => 'submit',
-      '#name' => 'context_' . $context_name,
-      '#attributes' => ['class' => ['rules-switch-button']],
-      '#parameter' => $context_name,
-      '#value' => $value,
-      '#submit' => ['::switchContextMode'],
-      // Do not validate!
-      '#limit_validation_errors' => [],
-    ];
-    return $form;
-  }
-
-  /**
-   * Submit callback: switch a context to data selecor or direct input mode.
-   */
-  public function switchContextMode(array &$form, FormStateInterface $form_state) {
-    $element_name = $form_state->getTriggeringElement()['#name'];
-    $mode = $form_state->get($element_name);
-    $switched_mode = $mode == 'selector' ? 'input' : 'selector';
-    $form_state->set($element_name, $switched_mode);
-
-    $form_state->setRebuild();
   }
 
   /**
